# swagger.yaml
swagger: "2.0"
info:
  description: "Atomist Explorer"
  version: 1.0.0
  title: Atomist Explorer
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "info@atomist.com"
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: /api/v1
...
schemes:
  - http
paths:
...
  /{workspace}/drift:
    get:
        tags:
          - drift
        summary: Show drift in the workspace
        description: Show drift in the workspace, in a tree structure where the outer ring is fingerprint
        operationId: drift
        produces:
          - application/json
        parameters:
          - name: workspace
            in: path
            description: Workspace id. * for all.
            required: true
            type: string
        parameters:
          - name: percentile
            in: query
            description: Entropy percentile (e.g. 80). Only show aspects with entropy above this.
            required: false
            type: integer
            format: int32
            default: 0
          - name: band
            in: query
            description: Introduce a ring for size band
            type: string
            enum:
              - true
              - false
            required: false
            default: false
        parameters:
          - name: type
            in: query
            description: Aspect name (fingerprint type) to show drift for
            required: false
            type: string
        responses:
          "200":
            description: successful operation
            schema:
              type: object
              properties:
                tree:
                  $ref: "#/definitions/Tree"
          "400":
            description: Invalid status value
...
definitions:
...
  Tree:
    properties:
      name:
        type: string
        description: Name of this level
      children:
        type: array
        items:
          # Probably need to do https://stackoverflow.com/questions/36866035/how-to-refer-to-enclosing-type-definition-recursively-in-openapi-swagger
          $ref: "#/definitions/Tree"
    required:
      - name